Query facet always shows when it has children and show more fix
Show more in query facet had a bug displaying it when not necessary.
This is now fixed by correctly computing when it needs to be displayed.

Query facets do not hide themselves entirely when no options are available but it has children.
It keeps the title bar and the children displayed.